refactor: dashboard layout, category dashboard layout
