+++ client/src/app/(dashboard)/(article)/article/[article]/page.tsx
