+++ client/src/app/(dashboard)/(categoryLayout)/category/[...label]/page.tsx
