+++ client/src/services/api/articleQuery.ts
@@ -1,4 +1,4 @@
-import { CreateArticleType } from '@/app/(article)/article/new-post/article.type';
+import { CreateArticleType } from '@/app/(dashboard)/(article)/article/new-post/article.type';
 import api, { AuthApi } from './base';
 import { IArticle } from '@/app/(dashboard)/types/articles.type';
 import { isAxiosError } from 'axios';
